Refactor: event 등록 함수 명칭 변경
