+++ public/component/Common/event/addHeaderEvent.js
@@ -1,6 +1,6 @@
 import Writing from '../../Writing/Writing.js';
 
-const addEventHeader = $parent => {
+const addHeaderEvent = $parent => {
   const $navUserWrapper = document.querySelector('.navbar-user-wrapper');
   const $navUserMenu = document.querySelector('.user-menu-list');
   const $navWritingBtn = document.querySelector('.writing-btn');
@@ -11,7 +11,7 @@ const addEventHeader = $parent => {
 
   $navWritingBtn.addEventListener('click', () => {
     window.history.pushState({}, '/writing', window.location.origin + '/writing');
-    $parent.lastElementChild.innerHTML = '';
+    $parent.removeChild($parent.lastChild);
     new Writing({
       $parent,
       initialState: {},
@@ -19,4 +19,4 @@ const addEventHeader = $parent => {
   });
 };
 
-export default addEventHeader;
+export default addHeaderEvent;
