+++ public/component/Writing/event/addWritingEvent.js
